split most nonrigid_optimization folder into new subfolders
