+++ src/nonrigid_optimization/hierarchical/hierarchical_optimizer2d.hpp
@@ -26,8 +26,8 @@
 #include <Eigen/Eigen>
 
 //local
-#include "../math/tensors.hpp"
-#include "../logging/convergence_report.hpp"
+#include "../../math/tensors.hpp"
+#include "../../logging/convergence_report.hpp"
 
 namespace eig = Eigen;
