+++ src/nonrigid_optimization/hierarchical/pyramid2d.hpp
