+++ src/nonrigid_optimization/slavcheva/data_term.cpp
@@ -16,10 +16,10 @@
 
 //local
 #include "data_term.hpp"
-#include "../math/tensors.hpp"
-#include "../math/boolean_operations.hpp"
-#include "../math/typedefs.hpp"
-#include "../traversal/field_traversal_cpu.hpp"
+#include "../../math/tensors.hpp"
+#include "../../math/boolean_operations.hpp"
+#include "../../math/typedefs.hpp"
+#include "../../traversal/field_traversal_cpu.hpp"
 
 namespace nonrigid_optimization {
