+++ src/nonrigid_optimization/slavcheva/full_gradient.hpp
